run extract script again on boost 1.47: now bcp wants more files
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@40226 a592a061-630c-0410-9148-cb99ea01b6c8

+++ boost/extract.sh
@@ -7,11 +7,19 @@
 #
 # Does also work with an outdated bcp version 
 #
+# Usage: extract.sh <path to new boost version>
+#
+
+if [ -z $1 ]
+then
+    echo "Usage: extract.sh <path to new boost version>"
+    exit 1
+fi
 
 rm -rf needed
 mkdir needed
 
-bcp --boost=$PWD \
+bcp --boost=$1 \
 	boost/any.hpp \
 	boost/assert.hpp \
 	boost/bind.hpp \
